[BEAM-3708] Simplifying precombine grouping tables.
Simplifying the grouping table code by making it more specific to the
precombine. The class doesn't need to be so generic when only the
Precombine runner is going to use it in the foreseeable future.
